Conflicts:
	kornell-api/src/main/scala/kornell/server/api/CourseClassResource.scala

+++ kornell-api/src/main/scala/kornell/server/jdbc/repository/RolesRepo.scala
@@ -62,6 +62,14 @@ class RolesRepo @Inject()(
 	      	| from Role r
 	        | where r.institution_uuid = ${institutionUUID}
 		    """.first[Role](toRole).get, RoleCategory.BIND_WITH_PERSON).getPerson
+		    
+  def getTutorsForCourseClass(courseClassUUID: String, bindMode: String)=
+      TOs.newRolesTO(sql"""
+            | select *
+            | from Role r
+            | where courseClassUUID = ${courseClassUUID}
+                | and r.role = ${RoleType.tutor.toString}
+            """.map[Role](toRole).map(bindRole(_, bindMode)))   
 
   private def bindRole(role: Role, bindMode: String) =
     TOs.newRoleTO(role, {
@@ -75,6 +83,8 @@ class RolesRepo @Inject()(
   def updateCourseClassAdmins(courseClassUUID: String, roles: Roles) = removeCourseClassAdmins(courseClassUUID).addRoles(roles)
   
   def updateInstitutionAdmins(institutionUUID: String, roles: Roles) = removeInstitutionAdmins(institutionUUID).addRoles(roles)
+  
+  def updateTutors(courseClassUUID: String, roles: Roles) = removeTutors(courseClassUUID).addRoles(roles)
   
   def addRoles(roles: Roles) = {
     roles.getRoles.asScala.foreach(addRole _)
@@ -106,6 +116,7 @@ class RolesRepo @Inject()(
     sql"""
     	delete from Role
     	where course_class_uuid = ${courseClassUUID}
+        and role = ${RoleType.courseClassAdmin}
     """.executeUpdate
     this
   }
@@ -114,6 +125,15 @@ class RolesRepo @Inject()(
     sql"""
         delete from Role
         where institution_uuid = ${institutionUUID}
+    """.executeUpdate
+    this
+  }
+  
+  def removeTutors(courseClassUUID: String) = {
+    sql"""
+        delete from Role
+        where course_class_uuid = ${courseClassUUID}
+        and role = ${RoleType.tutor}
     """.executeUpdate
     this
   }
